// src/typings/api/assets/assets.d.ts
export type RecordsListResp = {
  id: string;
  /** 币种 */
  businessCoin: string;
  /** 总金额 */
  total: string;
  /** 创建时间 */
  createdByTime: number;
  /** 状态，1、进行中 2、成功 3、失败 4、错误 */
  status: number;
  /** 类型，201. 充币，202. 提币 ,203 pay , 204. 冲正，501、现货手续费 502、合约手续费 503、提币手续费 504、锁仓手续费 */
  type: number;
  /** 合约类型 */
  groupType?: number;
  /** 手动/自动 */
  operationType?: string;
  /** 合约组名称 */
  groupName?: string;
  /** 划转来源 */
  from?: string;
  /** 划转去向 */
  to?: string;
  /** 订单 id */
  /** 返佣类型 */
  rebateTypeCd?: string;
  /** 现货订单 ID，现货买入卖出查询明细用 */
  orderId?: string;
}